Switch nncf.quantize to ptq for PT backend
